sustearn/requirements.txt
google-generativeai
python-dotenv
langchain
langchain_google_genai
django
djangorestframework
